fix: Long/Long integer division crash in AnalysisTab percentage
agg.totalAmount.toDouble() / totalExpense supaya hasil Double, bukan Long.
Tanpa ini, Long / Long = Long → String.format("%.1f") crash IllegalFormatConversionException.

+++ app/src/main/java/com/example/ui/analysis/AnalysisTab.kt
@@ -463,7 +463,7 @@ fun CombinedChartsCard(
                 Column(modifier = Modifier.fillMaxWidth()) {
                     aggregations.forEachIndexed { index, agg ->
                         val color = donutColors[index % donutColors.size]
-                        val percentage = if (totalExpense > 0) (agg.totalAmount / totalExpense) * 100 else 0.0
+                        val percentage = if (totalExpense > 0) (agg.totalAmount.toDouble() / totalExpense) * 100 else 0.0
 
                         Row(
                             verticalAlignment = Alignment.CenterVertically,
